+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added the `drivers/pgx` package which contains types and functions to use a `pgx/v5` connection as a `bob.Executor`.
 - Added support for `github.com/ncruces/go-sqlite3`. (thanks @ncruces)
+- Added support for configuring what type to generate specific null versions of a type instead of always wrapping with `sql.Null[T]`.
 
 ### Changed
 
@@ -19,8 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    This is to make it easier to implement for non `database/sql` drivers.
 - In the generated model, the `PrimaryKeyVals()` method is now private.
 - Renamed `driver_name` to `driver` in code generation configuration.
-- In type replacements, nullability is ONLY matched if explicitly set. Otherwise, the replacement applies based on other matches.  
-   Previously, nullability of matches was set to false by default, and would require an additional replacement to also match nullable columns.
+- In type replacements, nullability is not used in matching.
+  Previously, nullability of matches was set to false by default, and would require an additional replacement to also match nullable columns.
 
 ### Removed